requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==3.2
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.14.0
exceptiongroup==1.2.0
gunicorn==21.2.0
idna==3.6
iniconfig==2.0.0
Markdown==3.5.2
packaging==24.0
pluggy==1.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0
urllib3==2.2.1
validate-docbr==1.10.0
drf-yasg==1.21.7
